requirements.txt
Flask==2.2.1
mysql-connector-python==8.0.30
